Add an option to prevent speech being interrupted while scrolling the braille display (#13131)
Fixes #2124
# Summary of the issue:
Some people prefer to be able to read in braille while still listening to the speech output for the current line/object. This is currently not possible, because every braille gesture will interrupt speech.

# Description of user facing changes
Adds a new setting: "Interrupt speech while scrolling", default  is "enabled".
The new setting determines if speech should be interrupted when the Braille display is scrolled backwards/forwards.
Previous/next line commands are not affected.

This is added as an option because some speech+braille users prefer a bimodal interaction pattern.
- For these users, speech may be distracting while reading Braille.
- These users repetitively swap between primarily focusing on speech or braille.
- For these users, the current and expected UX for cancelling speech when reading braille is to scroll the line of braille.
- Considering alternatives:
  - Most braille devices cannot detect fingers reading the display.
  - Using an additional braille gesture just to cancel speech will slow users down / isn't a good use of a button on a display.

To preserve the existing behavior, the option is enabled by default, interrupting speech when scrolling.
- Disabling this option allows speech to be heard while the user is simultaneously reading Braille.
- Users who don't find speech distracting with Braille, or rely on both Braille and speech, may prefer to disable the option.

# Description of how this pull request fixes the issue:
Add an option to interrupt speech while the display is being scrolled.
This option is enabled by default, so the previous behavior is preserved.
This implementation only works for scroll previous/next display and does not affect other gestures such as previous/next line.