Request more specific information in bug report (#13905)
It would be helpful if bug reporters include exact speech/braille that is expected vs. occurring.
This is requested in devDocs\githubIssueTemplateExplanationAndExamples.md but not in the issue template.

It would also be good to add a section for logs and other attachments, as they are often required when reporting an issue.
This is also requested in devDocs\githubIssueTemplateExplanationAndExamples.md but not in the issue template.

Description of user facing changes
Bug reports are now prompted to include explicit speech/braille and NVDA logs.