Start a style guide for the User Guide (#13916)
Summary of the issue:
There is no set of standards or guide when writing documentation for the "User Guide".

Description of user facing changes
The intended audience for this PR is NVDA contributors.
NVDA developers now have suggested formatting for key commands and feature flags.

Feature flag documentation suggests using definition lists.
The default CSS formatting for definition lists is not ideal, the term and definition are listed on separate lines. There is no text formatting, the only indication that a term is linked to a definition is an indentation in front of the term.