Don't exit chrome after each test
Initially this was done to ensure a "clean slate" for each subsequent
test.
However, each test sample loads in a new tab already, the risk of
interference is low.

On the offchance that closing and opening chrome causes delays, or
background updates to run which may interfere with the tests on
appveyor, this is disabled.

For developers, it is easier to inspect test samples after the test
runs.
This may be important if the test fails.
However, it is noted that it leaves many tabs open for the developer to
tidy up.