Add new script to report accelerator key of currently focused object (#13960)
Summary of the issue:
There is no way to simply obtain the accelerator key of the currently focused object without reading its entire description.

Description of how this pull request fixes the issue:
Adds a new script to report the accelerator key. By default, it is bound to Shift+numpad 2 (desktop) and Ctrl+Shift+NVDA+. (laptop). As a mnemonic, accelerator keys are (or at least used to be) visually displayed by an underlined character.