Use line as a  unit for which say all is moved in rich edit controls (#13971)
Fixes #13420
Fixes #9179

Summary of the issue:
When in say all NVDA moves caret by the unit specified as a RREADINGCHUNK for a given text info implementation. In most text info implementations it defaults to line, however in ITextDocumentTextInfo it was set to sentence. While it probably results in smoother reading for literary text written in language where TOM can split into sentences properly, in text's such as log or for languages without clear concept of sentence such as Chinese it results in caret not following spoken text since it is moved to the current position only when the sentence ending is reached.

Description of user facing changes
Say all both for review and system cursor moves by line in Rich Edit controls such as WordPad or NVDA's log viewer.

Description of development approach
for ITextDocumentTextInfo textInfos.UNIT_READINGCHUNK has been set to line.