Add Quick Start Guide to NVDA User Guide (#13934)
Summary of the issue:
User requests for a more in-depth guide to help new users get started with NVDA. The guide has been written with the intention of including it in the User Guide, but also being able to easily separate it out into a standalone document where needed.

The quick start guide also helps bring users onboard in such a way as to give them enough information to get started, and also include information about other resources (eg training material and user groups), to ensure they are able to find additional support if needed.

Description of user facing changes
Move System Requirements to become section 1.2
Make new "Quick Start Guide" section 2, with sections on downloading NVDA, installing NVDA and Running NVDA (and smaller topics in the quick start guide on preferences, getting help, community.
The idea is that the Quick Start Guide will also be able to be used outside the user guide as a standalone document.
Section 3 "More setup options" includes most of the original section 3, but renamed as "Getting and setting up NVDA" could be confusing next to "Quick start guide". Most of the text is unchanged, as although slightly repetitious after the quick start guide, the quick start guide walks deeply through (default) options where "More setup options" covers other details, such as information on portable copies.
I appreciate this is a large documentation update and I apologise in advance to the translators. I am deeply grateful for all your work!

Description of development approach
Early work on this was included in the Switching from Jaws to NVDA Wiki some time ago and has been favourably commented on. The text has been drafted several times with the intention of complementing the remainder of the User Guide.