Track start and exit of notepad
Prior approach was flowed, the 'start.exe' process was being tracked
instead. The 'start.exe' process exited immediatly after starting notepad.
Now it waits until the process is complete.
Additionally the notepad RF lib has access to the Windows HWND allowing
for checking if the window is in the foreground.