Enable rdp via appveyor settings (#14004)
Related to #13983

Summary of the issue:
Sometimes the Appveyor environment is unpredictable, recently Docker Desktop opens a window asking for feedback. This window takes the foreground which subsequently results in applications started from the system test being unable to take focus back.

Description of user facing changes
For developers:
The intention is to fix this foreground issue separately, which may take some time. Until then, and for other issues in the future, make it easier to enable RDP as required.

The details of the connection show in the build log at the start of the build (after cloning NVDA)

Description of development approach
When the environment variable for an RDP password is set (via the Appveyor settings), the appveyor RDP script will be run.
NV Access developers will need to set (and unset) this environment variable as required.