Send architecture when update checking (#14019)
Summary of the issue:
It was discovered in #12064 that NVDA sends the architecture to the update server when updating, but it doesn't properly distinguish between architectures. Therefore, NVDA should send which architecture is in use specifically.

Description of user facing changes
None, only for usage stats.

Description of development approach
Consistency of data is kept by not changing how the x64 key is determined in the dictionary.
The specific architecture is added as an extra key to the dictionary.