Revert change of globalMapScripts in scriptHandler, which is now again a list instead of a set (#14066)
Fixes #14050

Summary of the issue:
Overriding/remapping existing braille display gestures was broken in NVDA 2022.2.1.

Description of user facing changes
Gesture overrides work again in the mentioned case in #14065

Description of development approach
Revert an uncessary and seemlinly undocumented change in scriptHandler. It would help if we could clear up how this change ended up in this release.