NVDAObjects.UIA: introduce UIA is grabbed property handler (#14097)
Closes #14081

Summary of the issue:
NVDA does not recognize if a UIA element is about to be dragged when using keyboards.

Description of user facing changes
When a UIA element such as Windows 10 Start menu tile is about to be dragged and rearranged, NVDA will say "dragging" to indicate the dragging state.

Description of development approach
Add UIA "IsGrabbed" property to list of states to be handled by UIA objects, and map drag start UIA event to state change event.

* NVDAObjects.UIA: add support for UIA 'IsGrabbed' property. Re #14081.

When elements are about to be dragged, UIA sets 'IsGrabed' property to True. Add this property to UIA object states set so it can be detected and announced by NVDA (announcement is next commit).

* UIA handler: map UIA is grabbed property event to state change event. Re #14081.

Treat 'is grabbed' property chnages as state changes. This will result in NVDA saying 'dragging' when items are grabbed, most notably when rearranging tiles in Windows 10 start menu.

* update changes

Co-authored-by: Sean Budd <sean@nvaccess.org>