UIA handler: add and map drag cancel and complete events to state change events. Re #14081. (#14104)
Additional fix for #14081
Follow-up to #14097

Summary of the issue:
NVDA is not announcing "dragging' when the item being dragged is let go and dragged again.

Description of user facing changes
When the item is grabbed a second tie, NVDA will announce 'dragging'.

Description of development approach
Add and map UIA drag cancel and complete events to state changes. This allows UIA object states cache to reflect the fact that "is grabbed" property is False, allowing NVDA to announce "dragging" when the item is grabbed again. At the moment NVDA will not say "done dragging" when drag complete event fires - that might be something to investigate in the future (remapping drag cancel/commplete events or doing soething with state change event/speech output).

Without defining UIA drag cancel and complete events, NVDA will not nanounce 'dragging' when dragging the selected item a second time (the added events do allow flushing staes cache as part of state change event, although NVDA will not nanounce 'done dragging' which is somethig to be investigated in the future).