Added Simplified Chinese Braille table 'Chinese (China, Mandarin) Current Braille System (no tones)' (#14138)
Summary of the issue:

zh-chn.ctb is a variant of the Chinese Current Braille System(zhcn-g1.ctb), which omits tones and allows for more braille content to be displayed for small braille displays.
In the nvda Chinese community, some people hope that this braille table can also be a candidate.
But there is no way to do this without manually renaming the table to replace the existing table.
So, I added it to braille settings dialog.

Description of user facing changes

Added Simplified Chinese Braille table

Chinese (China, Mandarin) Current Braille System (no tones)

Description of development approach

Add the definition of the table in brailleTables.py.