Make system tests pass when run locally on non-English systems (#14148)
Closes #13362

 Summary of the issue:
The majority of system tests fail when run on non-English systems. On my system (French), failures are due to the following reasons:
- Default NVDA's UI language is user's default Windows language, thus the messages of NVDA are reported in this language.
- If not specified, the language of an HTML page is set by the browser itself which may not be English (system language?)
- According to the locale keyboard layout, someemulated  keypress may fail. E.g. "NVDA+\" cannot be emulated with French keyboard layout since \ (backslash) can only be accessed with modifiers (AltGr).

In #13362, it was also stated that Chrome's language may make some tests fail. That's not the case on my French system. But we may imagine that the keyboard shortcuts could vary from one language to another, e.g. "Alt+d" to focus the address field.

 Description of user facing changes
System tests now pass on French system. It should pass on any system regardless of its language.

 Description of development approach

1. NVDA interface language is forced to English in the various config files.
2. HTML sample content forced to English (via lang HTML attribute)
3. Removed the unneeded NVDA+\ custom gesture which cannot be executed without modifier on some keyboard layout (e.g. French). Actually, there is no need for custom gesture since `NVDA+d` is now a native gesture of NVDA for the same script.
   Even if not used anymore for now, I have not removed the possibility to use a custom gesture file in case an unassigned script has to be tested in the future. Should it be the case, care should be taken to assign a gesture that can be executed on any keyboard layout, i.e. use a letter rather than punctuation as the main key name.
E. Chrome UI language is forced to English via command line parameter (see List of Chromium Command Line Switches « Peter Beverloo: https://peter.sh/experiments/chromium-command-line-switches/)

Note: For the Chrome command line --lang parameter to be taken into account, it should be ensured that no previous Chrome window is open when the tests are run.

---

* Make system tests pass when run locally on non-English systems.

For this, the following changes have been made:
- NVDA interface language forced to English (in config files)
- Chrome UI language forced to English (via command line parameter)
- HTML sample content forced to English (via lang HTML attribute)
- Removed the unneeded NVDA+\ custom gesture which cannot be executed without modifier on some keyboard layout (e.g. French)

Note: For the Chrome command line --lang parameter to be taken into account, it should be ensured that no previous Chrome window is open when the tests are run.

* Bump CI

* Remove unused .ini file.

* update changes

Co-authored-by: Sean Budd <sean@nvaccess.org>