Add support for embedded WPFTermControl in LTS Visual Studio (#14194)
Summary of the issue:
Visual Studio uses a WPF implementation of the Windows Terminal control that will use a different UIA class name. As such, NVDA currently does not add the needed overlay class to this control, so autoread and password suppression do not work as expected.

Description of how this pull request fixes the issue:
Added the new UIA class name to the necessary lists.