Report battery information first when fetching current status (#14215)
Fixes #14214
See also #14213

Summary of the issue:
The current battery status message when pressing NVDA+shift+b is too verbose, because it reports the AC status before the battery level.

This is useful for an AC state change (e.g. when plugging in a charger)
But is not ideal for when pressing NVDA+shift+b to fetch the battery status.

Description of user facing changes
Reverts to previous ordering in 2022.3.

When the AC status changes, the AC status is still reported first: "charging" or "AC disconnected".
When NVDA+shift+b is pressed to fetch the battery status, the battery status is reported first: "X percentage, Z hours and Y minutes remaining"

Description of development approach
Change internal parameter onlyReportIfStatusChanged to _ReportContext.
A _ReportContext is used to determine order of speech.

When the context is an AC status change, this reports the current AC status first.
When the context is a user fetching the current battery status, this reports the remaining battery life first.