When requesting formatting information, default colors are now explicitely reported in Wordpad and other edit fields (#13942)
Fixes #13959

Summary of the issue:
In Wordpad or in log viewer, pressing NVDA+F reports "default color on default color" as color information in formatting information. This information is of no use to know the color really displayed, what may be necessary when working with sighted persons.

In log viewer and other system edit fields, default colors are the ones provided by the system and can vary according to system settings, e.g. enabling high contrast theme.
In Wordpad, "Default color" is meant to be the one called "Automatic" and corresponds to system color.

Description of user facing changes
When pressing NVDA+F in Wordpad or NVDA log viewer, NVDA reports the text / background color and in addition, indicates if it is the default color, i.e. system color.
I have not bothered to report the application's color "Automatic" in Wordpad instead of "default" because it would probably need to create a WordPad appModule and specific processing just for this distinction; I think it was not worth it.

In MS Word, when the text color is set to "Automatic", "Automatic color" is reported instead of "default color". Indeed, this matches Word application terminology; in Word, automatic color and default color are two distinct concepts:
"automatic" is a color which adapts to favor a sufficient contrast with the background
"default color" is the color of the text when you open a new document (if registered in normal.dot) or when you remove formatting (when registered in the current document only).

Description of development approach
According to the CHARFORMATW structure for the class EditTextInfo and to ITextFont::GetForeColor for the class ITextDocumentTextInfo, the system colors are used for default colors. Thus, I retrieve these colors in these cases.