Merge pull request from GHSA-72mj-mqhj-qh4w
https://github.com/nvaccess/nvda/security/advisories/GHSA-72mj-mqhj-qh4w

Session tracking allows NVDA notifies of session lock state
changes. However, a transition is required to set the initial
state. This approach relied on the assumption that NVDA can't
start while already on the lock screen. While the assumption is
typically valid, it is possible (but unlikely) for NVDA to start
after the session is locked.

No user observable changes.

Initialize the lock state by querying the session info.