vbuf backend: If a node has been marked allowReuseInAncestorUpdate, ensure we propagate alwaysRerenderDescendants if appropriate before returning.
Previously, the check for allowReuseInAncestorUpdate was before alwaysRerenderDescendants.
This meant that if the node had allowReuseInAncestorUpdate and its parent had alwaysRerenderDescendants, we wouldn't propagate alwaysRerenderDescendants down the tree.
That meant that we might not pick up changes in a subtree which was moved at the same time as descendants were mutated.
This caused intermittent broken controls and other weirdness on GitHub issue pages.