Fixed an issue where symbol repetition was not correctly reported by the synthesizer. (#14266)
Summary of the issue:
Read the following line with OneCore (e.g. Zira) and symbol level all:
What is USD????

I can hear:
"What is four US dollars question"

Instead I expect
"What is USD four question"

The issue is due to the fact that the synth interprets "USD 4" as "four dollars". Instead, the repetition of "question" should not handled with the rest of the text by the synth but reported as is.

The impacted synthesizers are:

OneCore
SAPI5
IBMTTS
On the contrary, eSpeak is not impacted. Probably because it does not handle internally expressions such as "USD 4".

Description of user facing changes
The user should not hear anymore issues when reporting repeated symbols.

Description of development approach
Two spaces are added instead of only one between the rest of the text.